requirements.txt
pandas==2.1.3
numpy==1.26.2
aiosqlite==0.19.0
websockets==12.0
statsmodels==0.14.0
scipy==1.11.4
streamlit==1.29.0
plotly==5.18.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pyyaml==6.0.1
python-dateutil==2.8.2
pytz==2023.3